Another WindowFunction throw

+++ src/math/windowfunction.cpp
@@ -84,7 +84,7 @@ double WindowFunction::y(double x, double omega) const
              */
             return sin(x * (PI / xMax_)) / (x * (PI / xMax_));
         default:
-            throw(std::runtime_error(fmt::format("Window function enumeration {} not implemented.\n", form_)));
+            throw(std::runtime_error(fmt::format("Window function enumeration {} not implemented.\n", (int)form_)));
     }
 
     return 0.0;
